+++ R/read_tsca.R
@@ -8,17 +8,17 @@
 #'   format.
 #' @return Returns a data frame.
 #' @author Raoul Wolf (\url{https://github.com/RaoulWolf/})
-#' @note Tested with the "last created" version 2021-08.
+#' @note Tested with the "last created" version 2022-02.
 #' @examples \dontrun{
 #' download.file(
 #'   url = paste0(
-#'     "https://www.epa.gov/system/files/other-files/2021-08/",
-#'     "csv-non-cbi-tsca-inventory-202108.zip"
+#'     "https://www.epa.gov/system/files/other-files/2022-03/",
+#'     "csv-non-cbi-tsca-inventory-022022.zip"
 #'     ),
-#'   destfile = "csv-non-cbi-tsca-inventory-202108.zip"
+#'   destfile = "csv-non-cbi-tsca-inventory-022022.zip"
 #' )
 #'
-#' unzip(zipfile = "csv-non-cbi-tsca-inventory-202108.zip")
+#' unzip(zipfile = "csv-non-cbi-tsca-inventory-022022.zip")
 #'
 #' file_name <- list.files()
 #'
@@ -30,12 +30,24 @@
 #' @export
 read_tsca <- function(path, last_created = TRUE) {
 
-  tsca <- utils::read.csv(
-    file = path,
-    na.strings = c(""),
-    colClasses = c("UID" = "character", "EXP" = "integer"),
-    stringsAsFactors = FALSE
-  )
+  if (grepl(pattern = "2022", path)) {
+
+    tsca <- utils::read.csv(
+      file = path,
+      na.strings = c(""),
+      stringsAsFactors = FALSE
+    )
+
+  } else {
+
+    tsca <- utils::read.csv(
+      file = path,
+      na.strings = c(""),
+      colClasses = c("UID" = "character", "EXP" = "integer"),
+      stringsAsFactors = FALSE
+    )
+
+  }
 
   colnames(tsca) <- c(
     "id", "cas_rn", "cas_reg_no", "uid", "exp", "chem_name", "def", "uvcb",